% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Data from CSV File}
\usage{
read_data(
  file,
  path = NULL,
  colClasses = NA,
  create = TRUE,
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{file}{the name of the file to read. If path argument is not provided, it should contain the path to the file.}

\item{path}{the path where to read the file. If NULL (default), the file argument will be passed to read.csv().
Otherwise, it will be used to build the url (any path within file argument will then be ignored).}

\item{colClasses}{a vector of classes to be passed to read.csv(), or used to create the file and output value.
Possible values are (reminder from read.csv()):
\itemize{
\item NA (the default, when type.convert is used),
\item "NULL" (when the column is skipped),
\item one of the atomic vector classes (logical, integer, numeric, complex, character, raw),
\item or "factor", "Date" or "POSIXct"
Otherwise there needs to be an as method (from package methods) for conversion from "character" to the specified formal class.
}}

\item{create}{a logical value (default = TRUE) to indicate whether non existing file and/or path should be created or not.
check create_data() functions to get details about different behaviors.}

\item{encoding}{a character value (default = "UTF8"), passed to read.csv().}

\item{...}{any argument you want to pass to read.csv() or write_data() if called.}
}
\value{
a data.frame, output of read.csv().
}
\description{
Read Data from CSV File
}
\details{
If create is TRUE and colClasses is NA, then file creation will be skipped as an empty data.frame (0 obs. of 0 variables)
will be received from create_data()
}
\examples{
\dontrun{read_data(file = "my_file.csv",
path = "./some/folder",
colClasses = c("id" = "numeric", "date" = "Date", "comment" = "character"))}

}
