% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Write Data to CSV File}
\usage{
write_data(
  data,
  file,
  path = NULL,
  create = TRUE,
  row.names = FALSE,
  fileEncoding = "UTF-8",
  ...
)
}
\arguments{
\item{data}{a data.frame of the data to be saved.}

\item{file}{the name of the file to read. If path argument is not provided, it should contain the path to the file.}

\item{path}{the path where to read the file. If NULL (default), the file argument will be passed to write.csv().
Otherwise, it will be used to build the url (any path within file argument will then be ignored).}

\item{create}{a logical value (default = TRUE) to indicate whether non existing path should be created or not.}

\item{row.names}{a logical value (default = FALSE) to indicate whether row names should be written or not.}

\item{fileEncoding}{a character value (default = "UTF8"), passed to write.csv().}

\item{...}{any argument you want to pass to write.csv().}
}
\value{
FALSE if path (either the path argument or dirname of file) does not exist and create is FALSE
}
\description{
Write Data to CSV File
}
\examples{
\dontrun{write_data(data.frame(id = 1, comment = "some text"),
file = "my_file.csv",
path = "./some/folder",
create = TRUE,
row.names = FALSE,
fileEncoding = "UTF-8")}
}
